tests: split put_get_remove test
Basic version only tests pmemkv with key length <= 8 bytes.
